Return correct size from b64_decode.
The current implementation does not take into account padding (and yet it does
actually depend on padding), or skipped whitespace/gibberish characters.

The actual size can be computed by subtracting the padding and gibberish bytes,
but in this case it's easier to just return the actual number of bytes written.